add new sharable display trait, separate from shared display
